style: rename file from pr comments
