+++ src/pages/collection/[collectionId]/index.tsx
@@ -1,5 +1,5 @@
 import CollectionDetails from '@app/components/CollectionDetails';
-import { getAuthHeaders } from '@app/utils/localRequestHelper';
+import { getAuthHeaders } from '@app/utils/serverSidePropsHelpers';
 import type { Collection } from '@server/models/Collection';
 import type { GetServerSideProps, NextPage } from 'next';
